// sanity.types.ts
export type Address = {
  _id: string;
  _type: "address";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  fullName?: string;
  phone?: string;
  division?:
    | "Dhaka"
    | "Chittagong"
    | "Rajshahi"
    | "Khulna"
    | "Barisal"
    | "Sylhet"
    | "Rangpur"
    | "Mymensingh";
  district?: string;
  upazila?: string;
  postcode?: string;
  streetAddress?: string;
  userId?: string;
};